Restructured and reformatted
